Use floats and sinusoidal vertical motion
Convert Ball members to float types and replace linear Y movement with a sinusoidal vertical motion. Initialize speedX with a random distribution, add RNG setup, and tweak initial position. Implement X boundary handling, range decay, and speed decay; add mouse input to boost horizontal speed and range. Draw debug text for position and speeds. Also increase target FPS from 60 to 144 in main.cpp.

+++ Raylib-CPP-Starter-Template-for-VSCODE-V2-main/src/ball.h
@@ -8,9 +8,9 @@ public:
     void Draw() const;
 
 private:
-    int x;
-    int y;
-    int speedX;
-    int speedY;
-    int radius;
+    float x;
+    float y;
+    float speedX;
+    float speedY;
+    float radius;
 };
